Add options_for_association_joins
association_options_find and association_options_count only supported conditions, but a lot of conditions needed joins to make sense.  To work around not having joins, a separate full query with a join could be performed and then the id's from this query used as the condition, but it was inefficient, so add options_for_association_joins to go with options_for_association_conditions to add any necessary joins for the conditons to work.