Generic onchange / update column support
The onchange support previously only worked for singular values, such as select dropdown.  This change adds a form enhancement, Form.columnValue(form, input_name) to serialize all form elements that have names that start with the same input_name.  The form enchance means that the onchange works on the select and on the checkboxs that make up part of a checkbox group. Since no checkbox could be checked, and therefore, an empty serialization could be sent, params[:record] and params[:record][column.name] are checked for nil.  params[:value] is no longer supported.