Pass link to security method if it will accept it
Check arity of link security method in skip_action_link.  If arity is 1 then pass the link to the method, otherwise use the old behavior of calling the security method without a
rguments.  Passing the link allows for security logic based on where the link is pointing, which can be useful for nested_authorized? as links to associations that the user can'
t read can then be skipped.  Determining which link nested_authorized was authorizing would be very difficult without this change.